extract the portfolio form logic to a separate component
